+++ relay-config/src/aggregator.rs
@@ -165,6 +165,6 @@ mod tests {
         let condition = serde_json::from_value::<Condition>(json).unwrap();
         assert!(condition.matches(Some(MetricNamespace::Spans)));
         assert!(condition.matches(Some(MetricNamespace::Custom)));
-        assert!(!condition.matches(Some(MetricNamespace::Transactions)));
+        assert!(!condition.matches(Some(MetricNamespace::Sessions)));
     }
 }