+++ relay-metrics/src/bucket.rs
@@ -927,15 +927,15 @@ mod tests {
 
     #[test]
     fn test_parse_counter() {
-        let s = "transactions/foo:42|c";
+        let s = "spans/foo:42|c";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "c:transactions/foo@none",
+                "c:spans/foo@none",
             ),
             value: Counter(
                 42.0,
@@ -947,12 +947,12 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
     fn test_parse_counter_packed() {
-        let s = "transactions/foo:42:17:21|c";
+        let s = "spans/foo:42:17:21|c";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
         assert_eq!(metric.value, BucketValue::Counter(80.into()));
@@ -960,15 +960,15 @@ mod tests {
 
     #[test]
     fn test_parse_distribution() {
-        let s = "transactions/foo:17.5|d";
+        let s = "spans/foo:17.5|d";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "d:transactions/foo@none",
+                "d:spans/foo@none",
             ),
             value: Distribution(
                 [
@@ -982,7 +982,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -1013,15 +1013,15 @@ mod tests {
 
     #[test]
     fn test_parse_set() {
-        let s = "transactions/foo:4267882815|s";
+        let s = "spans/foo:4267882815|s";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "s:transactions/foo@none",
+                "s:spans/foo@none",
             ),
             value: Set(
                 {
@@ -1035,7 +1035,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -1070,15 +1070,15 @@ mod tests {
 
     #[test]
     fn test_parse_gauge() {
-        let s = "transactions/foo:42|g";
+        let s = "spans/foo:42|g";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "g:transactions/foo@none",
+                "g:spans/foo@none",
             ),
             value: Gauge(
                 GaugeValue {
@@ -1096,20 +1096,20 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
     fn test_parse_gauge_packed() {
-        let s = "transactions/foo:25:17:42:220:85|g";
+        let s = "spans/foo:25:17:42:220:85|g";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "g:transactions/foo@none",
+                "g:spans/foo@none",
             ),
             value: Gauge(
                 GaugeValue {
@@ -1127,7 +1127,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]