+++ tests/integration/test_cardinality_limiter.py
@@ -37,11 +37,11 @@ def test_cardinality_limits(mini_sentry, relay_with_processing, metrics_consumer
     project_id = 10000
     cardinality_limits = [
         {
-            "id": "transactions",
+            "id": "spans",
             "window": {"windowSeconds": 3600, "granularitySeconds": 600},
             "limit": 1,
             "scope": "project",
-            "namespace": "transactions",
+            "namespace": "spans",
         },
         {
             "id": "custom",
@@ -56,8 +56,8 @@ def test_cardinality_limits(mini_sentry, relay_with_processing, metrics_consumer
 
     metrics_payload = "\n".join(
         [
-            "transactions/foo@second:12|c",
-            "transactions/bar@second:23|c",
+            "spans/foo@second:12|c",
+            "spans/bar@second:23|c",
             "sessions/foo@second:12|c",
             "foo@second:12|c",
             "bar@second:23|c",
@@ -69,7 +69,7 @@ def test_cardinality_limits(mini_sentry, relay_with_processing, metrics_consumer
     metrics = metrics_by_namespace(metrics_consumer, 4)
     assert len(metrics["custom"]) == 2
     assert len(metrics["sessions"]) == 1
-    assert len(metrics["transactions"]) == 1
+    assert len(metrics["spans"]) == 1
 
 
 @pytest.mark.parametrize("mode", [None, "enabled", "disabled", "passive"])
@@ -85,25 +85,25 @@ def test_cardinality_limits_global_config_mode(
     project_id = 10001
     cardinality_limits = [
         {
-            "id": "transactions",
+            "id": "spans",
             "window": {"windowSeconds": 3600, "granularitySeconds": 600},
             "limit": 1,
             "scope": "project",
-            "namespace": "transactions",
+            "namespace": "spans",
         },
     ]
 
     add_project_config(mini_sentry, project_id, cardinality_limits)
 
-    metrics_payload = "transactions/foo@second:12|c\ntransactions/bar@second:23|c"
+    metrics_payload = "spans/foo@second:12|c\nspans/bar@second:23|c"
     relay.send_metrics(project_id, metrics_payload)
 
     if mode in [None, "enabled"]:
         metrics = metrics_by_namespace(metrics_consumer, 1)
-        assert len(metrics["transactions"]) == 1
+        assert len(metrics["spans"]) == 1
     else:
         metrics = metrics_by_namespace(metrics_consumer, 2)
-        assert len(metrics["transactions"]) == 2
+        assert len(metrics["spans"]) == 2
 
 
 def test_cardinality_limits_passive_limit(
@@ -115,19 +115,19 @@ def test_cardinality_limits_passive_limit(
     project_id = 10002
     cardinality_limits = [
         {
-            "id": "transactions_passive",
+            "id": "spans_passive",
             "passive": True,
             "window": {"windowSeconds": 3600, "granularitySeconds": 600},
             "limit": 1,
             "scope": "project",
-            "namespace": "transactions",
+            "namespace": "spans",
         },
         {
-            "id": "transactions_enforced",
+            "id": "spans_enforced",
             "window": {"windowSeconds": 3600, "granularitySeconds": 600},
             "limit": 3,
             "scope": "project",
-            "namespace": "transactions",
+            "namespace": "spans",
         },
         {
             "id": "custom",
@@ -142,10 +142,10 @@ def test_cardinality_limits_passive_limit(
 
     metrics_payload = "\n".join(
         [
-            "transactions/foo@second:11|c",
-            "transactions/bar@second:22|c",
-            "transactions/baz@second:33|c",
-            "transactions/lol@second:55|c",
+            "spans/foo@second:11|c",
+            "spans/bar@second:22|c",
+            "spans/baz@second:33|c",
+            "spans/lol@second:55|c",
             "sessions/foo@second:12|c",
             "foo@second:12|c",
             "bar@second:23|c",
@@ -158,4 +158,4 @@ def test_cardinality_limits_passive_limit(
     assert len(metrics["custom"]) == 2
     assert len(metrics["sessions"]) == 1
     # The passive limit should be ignored, the non-passive limit still needs to be enforced.
-    assert len(metrics["transactions"]) == 3
+    assert len(metrics["spans"]) == 3